Update quizOper.js
Added the skipQuestion() function. If a question in the quiz may be skipped, the quiz author could add a button to call skipQuestion() from that question. Then skipQuestion will skip ahead to the next question and update the variable skippedQuestionsCount so that at the end the number of skipped questions is deducted from the total possible score.