Update quizStyle.css
Since some people may be red-green color blind, it is necessary to provide some other cue that a particular answer to the quiz is right or wrong. This is accomplished with the CSS :before property. Also, since the quiz author may wish to elaborate on a correct answer almost as much as a wrong answer, .afterAnswerInfo is modified so that it can be in either a right answer or a wrong answer.